+++ src/tilda/Gen.java
@@ -23,11 +23,11 @@ import org.apache.logging.log4j.Logger;
 
 import tilda.generation.Generator;
 import tilda.generation.GeneratorSession;
+import tilda.generation.GraphvizUtil;
 import tilda.generation.Manifest;
 import tilda.parsing.ParserSession;
 import tilda.parsing.Parser;
 import tilda.parsing.parts.Schema;
-import tilda.utils.GraphvizUtil;
 import tilda.utils.SystemValues;
 
 public class Gen
